Move over metadata fetching functions
From solr_builder to parent class DataProvider
