+++ openlibrary/solr/data_provider.py
@@ -5,9 +5,14 @@ data required for solr.
 
 Multiple data providers are supported, each is good for different use case.
 """
+import asyncio
+import itertools
 import logging
-from typing import Dict, List, Optional
+import re
+from typing import Iterable, List, Optional, Sized
 
+import httpx
+from httpx import HTTPError
 import requests
 import web
 from web import DB
@@ -18,6 +23,7 @@ from openlibrary.core import ia
 logger = logging.getLogger("openlibrary.solr.data_provider")
 
 IA_METADATA_FIELDS = ('identifier', 'boxid', 'collection')
+OCAID_PATTERN = re.compile(r'^[^\s&#?/]+$')
 
 
 def get_data_provider(type="default"):
@@ -26,6 +32,80 @@ def get_data_provider(type="default"):
         return BetterDataProvider()
     elif type == "legacy":
         return LegacyDataProvider()
+
+
+def is_valid_ocaid(ocaid: str):
+    return bool(OCAID_PATTERN.match(ocaid))
+
+
+def batch(items: list, max_batch_len: int):
+    """
+    >>> list(batch([1,2,3,4,5], 2))
+    [[1, 2], [3, 4], [5]]
+    >>> list(batch([], 2))
+    []
+    >>> list(batch([1,2,3,4,5], 3))
+    [[1, 2, 3], [4, 5]]
+    >>> list(batch([1,2,3,4,5], 5))
+    [[1, 2, 3, 4, 5]]
+    >>> list(batch([1,2,3,4,5], 6))
+    [[1, 2, 3, 4, 5]]
+    """
+    start = 0
+    while start < len(items):
+        yield items[start : start + max_batch_len]
+        start += max_batch_len
+
+
+def batch_until_len(items: Iterable[Sized], max_batch_len: int):
+    batch_len = 0
+    batch = []  # type: List[Sized]
+    for item in items:
+        if batch_len + len(item) > max_batch_len and batch:
+            yield batch
+            batch = [item]
+            batch_len = len(item)
+        else:
+            batch.append(item)
+            batch_len += len(item)
+    if batch:
+        yield batch
+
+
+def partition(lst: list, parts: int):
+    """
+    >>> list(partition([1,2,3,4,5,6], 1))
+    [[1, 2, 3, 4, 5, 6]]
+    >>> list(partition([1,2,3,4,5,6], 2))
+    [[1, 2, 3], [4, 5, 6]]
+    >>> list(partition([1,2,3,4,5,6], 3))
+    [[1, 2], [3, 4], [5, 6]]
+    >>> list(partition([1,2,3,4,5,6], 4))
+    [[1], [2], [3], [4, 5, 6]]
+    >>> list(partition([1,2,3,4,5,6], 5))
+    [[1], [2], [3], [4], [5, 6]]
+    >>> list(partition([1,2,3,4,5,6], 6))
+    [[1], [2], [3], [4], [5], [6]]
+    >>> list(partition([1,2,3,4,5,6], 7))
+    [[1], [2], [3], [4], [5], [6]]
+
+    >>> list(partition([1,2,3,4,5,6,7], 3))
+    [[1, 2], [3, 4], [5, 6, 7]]
+
+    >>> list(partition([], 5))
+    []
+    """
+    if not lst:
+        return
+
+    total_len = len(lst)
+    parts = min(total_len, parts)
+    size = total_len // parts
+
+    for i in range(0, parts):
+        start = i * size
+        end = total_len if (i == parts - 1) else ((i + 1) * size)
+        yield lst[start:end]
 
 
 class DataProvider:
@@ -37,6 +117,82 @@ class DataProvider:
     in this module.
     """
 
+    def __init__(self) -> None:
+        self.ia_cache: dict[str, Optional[dict]] = dict()
+
+    @staticmethod
+    async def _get_lite_metadata(ocaids: list[str], _recur_depth=0, _max_recur_depth=3):
+        """
+        For bulk fetch, some of the ocaids in Open Library may be bad
+        and break archive.org ES fetches. When this happens, we (up to
+        3 times) recursively split up the pool of ocaids to do as many
+        successful sub-bulk fetches as we can and then when limit is
+        reached, downstream code will fetch remaining ocaids individually
+        (and skip bad ocaids)
+        """
+        if not ocaids or _recur_depth > _max_recur_depth:
+            logger.warning(
+                'Max recursion exceeded trying fetch IA data', extra={'ocaids': ocaids}
+            )
+            return []
+
+        try:
+            async with httpx.AsyncClient() as client:
+                r = await client.get(
+                    "https://archive.org/advancedsearch.php",
+                    timeout=30,  # The default is silly short
+                    params={
+                        'q': f"identifier:({' OR '.join(ocaids)})",
+                        'rows': len(ocaids),
+                        'fl': ','.join(IA_METADATA_FIELDS),
+                        'page': 1,
+                        'output': 'json',
+                        'save': 'yes',
+                    },
+                )
+            r.raise_for_status()
+            return r.json()['response']['docs']
+        except HTTPError:
+            logger.warning("IA bulk query failed")
+        except (ValueError, KeyError):
+            logger.warning(f"IA bulk query failed {r.status_code}: {r.json()['error']}")
+
+        # Only here if an exception occurred
+        # there's probably a bad apple; try splitting the batch
+        parts = await asyncio.gather(
+            *(
+                DataProvider._get_lite_metadata(part, _recur_depth=_recur_depth + 1)
+                for part in partition(ocaids, 6)
+            )
+        )
+        return list(itertools.chain(*parts))
+
+    @staticmethod
+    async def _get_lite_metadata_direct(ocaid: str):
+        try:
+            async with httpx.AsyncClient() as client:
+                r = await client.get(
+                    f"https://archive.org/metadata/{ocaid}/metadata",
+                    timeout=30,  # The default is silly short
+                )
+            r.raise_for_status()
+            response = r.json()
+            if 'error' not in response:
+                lite_metadata = {
+                    key: response['result'][key]
+                    for key in IA_METADATA_FIELDS
+                    if key in response['result']
+                }
+                return lite_metadata
+            else:
+                return {
+                    'error': response['error'],
+                    'identifier': ocaid,
+                }
+        except HTTPError:
+            logger.warning(f'Error fetching metadata for {ocaid}')
+            return None
+
     def get_document(self, key):
         """Returns the document with specified key from the database.
 
@@ -45,15 +201,15 @@ class DataProvider:
         """
         raise NotImplementedError()
 
-    def get_metadata(self, identifier):
-        """
-        Returns archive.org metadata for given identifier.
-
-        :param str identifier: Internet Archive id (aka ocaid)
-        :return:
-        :rtype: web.storage or None
-        """
-        raise NotImplementedError()
+    def get_metadata(self, identifier: str):
+        if identifier in self.ia_cache:
+            logger.debug("IA metadata cache hit")
+            return self.ia_cache[identifier]
+        elif not is_valid_ocaid(identifier):
+            return None
+        else:
+            logger.debug("IA metadata cache miss")
+            return ia.get_metadata_direct(identifier)
 
     def preload_documents(self, keys):
         """
@@ -65,12 +221,30 @@ class DataProvider:
         """
         pass
 
-    def preload_metadata(self, identifiers):
-        """
-        :param list of str identifiers: list of Internet Archive ids (aka ocaids)
-        :return:
-        """
-        pass
+    async def preload_metadata(self, ocaids: list[str]):
+        invalid_ocaids = {ocaid for ocaid in ocaids if not is_valid_ocaid(ocaid)}
+        if invalid_ocaids:
+            logger.warning(f"Trying to cache invalid OCAIDs: {invalid_ocaids}")
+        valid_ocaids = list(set(ocaids) - invalid_ocaids)
+        batches = list(batch_until_len(valid_ocaids, 3000))
+        # Start them all async
+        tasks = [asyncio.create_task(self._get_lite_metadata(b)) for b in batches]
+        for task in tasks:
+            for doc in await task:
+                self.ia_cache[doc['identifier']] = doc
+
+        missing_ocaids = [ocaid for ocaid in valid_ocaids if ocaid not in self.ia_cache]
+        missing_ocaid_batches = list(batch(missing_ocaids, 6))
+        for ocaids in missing_ocaid_batches:
+            # Start them all async
+            tasks = [
+                asyncio.create_task(self._get_lite_metadata_direct(ocaid))
+                for ocaid in ocaids
+            ]
+            for task in tasks:
+                lite_metadata = await task
+                if lite_metadata:
+                    self.ia_cache[lite_metadata['identifier']] = lite_metadata
 
     def preload_editions_of_works(self, work_keys):
         """
@@ -100,13 +274,14 @@ class DataProvider:
         raise NotImplementedError()
 
     def clear_cache(self):
-        raise NotImplementedError()
+        self.ia_cache.clear()
 
 
 class LegacyDataProvider(DataProvider):
     def __init__(self):
         from openlibrary.catalog.utils.query import query_iter, withKey
 
+        super().__init__()
         self._query_iter = query_iter
         self._withKey = withKey
 
@@ -121,17 +296,9 @@ class LegacyDataProvider(DataProvider):
         q = {'type': '/type/edition', 'works': work['key'], '*': None}
         return list(self._query_iter(q))
 
-    def get_metadata(self, identifier):
-        logger.info("find_metadata %s", identifier)
-        return ia.get_metadata(identifier)
-
     def get_document(self, key):
         logger.info("get_document %s", key)
         return self._withKey(key)
-
-    def clear_cache(self):
-        # Nothing's cached, so nothing to clear!
-        return
 
 
 class ExternalDataProvider(DataProvider):
@@ -140,6 +307,7 @@ class ExternalDataProvider(DataProvider):
     """
 
     def __init__(self, ol_host: str):
+        super().__init__()
         self.ol_host = ol_host
 
     def find_redirects(self, key: str):
@@ -154,15 +322,8 @@ class ExternalDataProvider(DataProvider):
             logger.warning(f"Too many editions for {work['key']}")
         return resp['entries']
 
-    def get_metadata(self, identifier: str):
-        return ia.get_metadata(identifier)
-
     def get_document(self, key: str):
         return requests.get(f"http://{self.ol_host}{key}.json").json()
-
-    def clear_cache(self):
-        # Nothing's cached, so nothing to clear!
-        return
 
 
 class BetterDataProvider(LegacyDataProvider):
@@ -178,11 +339,10 @@ class BetterDataProvider(LegacyDataProvider):
         infogami._setup()
         from infogami import config
         """
-        LegacyDataProvider.__init__(self)
+        super().__init__()
 
         # cache for documents
         self.cache: dict[str, dict] = {}
-        self.metadata_cache: dict[str, Optional[dict]] = {}
 
         # cache for redirects
         self.redirect_cache: dict[str, list[str]] = {}
@@ -204,51 +364,6 @@ class BetterDataProvider(LegacyDataProvider):
             self.db: DB = get_db()
         else:
             self.db = db
-
-    def get_metadata(self, identifier):
-        """Alternate implementation of ia.get_metadata() that uses ES."""
-        logger.info("get_metadata %s", identifier)
-        # ES will fail for no-scope items, we should have
-        # a role-user (creds) which accesses noindex
-        self.preload_metadata([identifier])
-        if not self.metadata_cache.get(identifier):
-            md = {
-                k: v
-                for (k, v) in ia.get_metadata(identifier).items()
-                if k in IA_METADATA_FIELDS
-            }
-            self.metadata_cache[identifier] = md
-        return web.storage(self.metadata_cache[identifier])
-
-    def preload_metadata(self, identifiers: list[str]):
-        identifiers = [id for id in identifiers if id not in self.metadata_cache]
-
-        logger.info("preload_metadata %s", identifiers)
-
-        if not identifiers:
-            return
-
-        r = requests.get(
-            "https://archive.org/advancedsearch.php",
-            params={
-                'q': f"identifier:({' OR '.join(identifiers)})",
-                'rows': str(len(identifiers)),
-                'fl': ','.join(IA_METADATA_FIELDS),
-                'page': '1',
-                'output': 'json',
-                'save': 'yes',
-            },
-        )
-
-        if r.status_code == 200:
-            print(r.json())
-            rows = r.json()['response']['docs']
-            for row in rows:
-                row = web.storage(row)
-                self.metadata_cache[row.identifier] = row
-
-        for identifier in identifiers:
-            self.metadata_cache.setdefault(identifier, None)
 
     def get_document(self, key):
         # logger.info("get_document %s", key)
@@ -391,7 +506,7 @@ class BetterDataProvider(LegacyDataProvider):
         return
 
     def clear_cache(self):
+        super().clear_cache()
         self.cache.clear()
-        self.metadata_cache.clear()
         self.redirect_cache.clear()
         self.edition_keys_of_works_cache.clear()
