+++ scripts/solr_builder/solr_builder/solr_builder.py
@@ -1,11 +1,7 @@
-import asyncio
-import itertools
 import json
-import re
-from typing import Awaitable, List, Iterable, Literal, Sized
+from typing import Awaitable, Literal
 
-import httpx
-from httpx import HTTPError
+
 from configparser import ConfigParser
 import logging
 import time
@@ -14,17 +10,12 @@ from collections import namedtuple
 
 import psycopg2
 
-from openlibrary.core import ia
 from openlibrary.solr import update_work
-from openlibrary.solr.data_provider import IA_METADATA_FIELDS, DataProvider
+from openlibrary.solr.data_provider import DataProvider
 from openlibrary.solr.update_work import load_configs, update_keys
 
+
 logger = logging.getLogger("openlibrary.solr-builder")
-OCAID_PATTERN = re.compile(r'^[^\s&#?/]+$')
-
-
-def is_valid_ocaid(ocaid: str):
-    return bool(OCAID_PATTERN.match(ocaid))
 
 
 def config_section_to_dict(config_file, section):
@@ -40,76 +31,6 @@ def config_section_to_dict(config_file, section):
     config.read(config_file)
     result = {key: config.get(section, key) for key in config.options(section)}
     return result
-
-
-def partition(lst: list, parts: int):
-    """
-    >>> list(partition([1,2,3,4,5,6], 1))
-    [[1, 2, 3, 4, 5, 6]]
-    >>> list(partition([1,2,3,4,5,6], 2))
-    [[1, 2, 3], [4, 5, 6]]
-    >>> list(partition([1,2,3,4,5,6], 3))
-    [[1, 2], [3, 4], [5, 6]]
-    >>> list(partition([1,2,3,4,5,6], 4))
-    [[1], [2], [3], [4, 5, 6]]
-    >>> list(partition([1,2,3,4,5,6], 5))
-    [[1], [2], [3], [4], [5, 6]]
-    >>> list(partition([1,2,3,4,5,6], 6))
-    [[1], [2], [3], [4], [5], [6]]
-    >>> list(partition([1,2,3,4,5,6], 7))
-    [[1], [2], [3], [4], [5], [6]]
-
-    >>> list(partition([1,2,3,4,5,6,7], 3))
-    [[1, 2], [3, 4], [5, 6, 7]]
-
-    >>> list(partition([], 5))
-    []
-    """
-    if not lst:
-        return
-
-    total_len = len(lst)
-    parts = min(total_len, parts)
-    size = total_len // parts
-
-    for i in range(0, parts):
-        start = i * size
-        end = total_len if (i == parts - 1) else ((i + 1) * size)
-        yield lst[start:end]
-
-
-def batch_until_len(items: Iterable[Sized], max_batch_len: int):
-    batch_len = 0
-    batch = []  # type: List[Sized]
-    for item in items:
-        if batch_len + len(item) > max_batch_len and batch:
-            yield batch
-            batch = [item]
-            batch_len = len(item)
-        else:
-            batch.append(item)
-            batch_len += len(item)
-    if batch:
-        yield batch
-
-
-def batch(items: list, max_batch_len: int):
-    """
-    >>> list(batch([1,2,3,4,5], 2))
-    [[1, 2], [3, 4], [5]]
-    >>> list(batch([], 2))
-    []
-    >>> list(batch([1,2,3,4,5], 3))
-    [[1, 2, 3], [4, 5]]
-    >>> list(batch([1,2,3,4,5], 5))
-    [[1, 2, 3, 4, 5]]
-    >>> list(batch([1,2,3,4,5], 6))
-    [[1, 2, 3, 4, 5]]
-    """
-    start = 0
-    while start < len(items):
-        yield items[start : start + max_batch_len]
-        start += max_batch_len
 
 
 def safeget(func):
@@ -137,12 +58,11 @@ class LocalPostgresDataProvider(DataProvider):
         """
         :param str db_conf_file: file to DB config with [postgres] section
         """
+        super().__init__()
         self._db_conf = config_section_to_dict(db_conf_file, "postgres")
         self._conn = None  # type: psycopg2._psycopg.connection
         self.cache = dict()
         self.cached_work_editions_ranges = []
-        self.ia_cache = dict()
-        self.ia = None
 
     def __enter__(self):
         """
@@ -213,106 +133,6 @@ class LocalPostgresDataProvider(DataProvider):
 
         cur.close()
 
-    @staticmethod
-    async def _get_lite_metadata(ocaids, _recur_depth=0, _max_recur_depth=3):
-        """
-        For bulk fetch, some of the ocaids in Open Library may be bad
-        and break archive.org ES fetches. When this happens, we (up to
-        3 times) recursively split up the pool of ocaids to do as many
-        successful sub-bulk fetches as we can and then when limit is
-        reached, downstream code will fetch remaining ocaids individually
-        (and skip bad ocaids)
-        """
-        if not ocaids or _recur_depth > _max_recur_depth:
-            logger.warning(
-                'Max recursion exceeded trying fetch IA data', extra={'ocaids': ocaids}
-            )
-            return []
-
-        try:
-            async with httpx.AsyncClient() as client:
-                r = await client.get(
-                    "https://archive.org/advancedsearch.php",
-                    timeout=30,  # The default is silly short
-                    params={
-                        'q': f"identifier:({' OR '.join(ocaids)})",
-                        'rows': len(ocaids),
-                        'fl': ','.join(IA_METADATA_FIELDS),
-                        'page': 1,
-                        'output': 'json',
-                        'save': 'yes',
-                    },
-                )
-            r.raise_for_status()
-            return r.json()['response']['docs']
-        except HTTPError:
-            logger.warning("IA bulk query failed")
-        except (ValueError, KeyError):
-            logger.warning(f"IA bulk query failed {r.status_code}: {r.json()['error']}")
-
-        # Only here if an exception occurred
-        # there's probably a bad apple; try splitting the batch
-        parts = await asyncio.gather(
-            *(
-                LocalPostgresDataProvider._get_lite_metadata(
-                    part, _recur_depth=_recur_depth + 1
-                )
-                for part in partition(ocaids, 6)
-            )
-        )
-        return list(itertools.chain(*parts))
-
-    @staticmethod
-    async def _get_lite_metadata_direct(ocaid: str):
-        try:
-            async with httpx.AsyncClient() as client:
-                r = await client.get(
-                    f"https://archive.org/metadata/{ocaid}/metadata",
-                    timeout=30,  # The default is silly short
-                )
-            r.raise_for_status()
-            response = r.json()
-            if 'error' not in response:
-                lite_metadata = {
-                    key: response['result'][key]
-                    for key in IA_METADATA_FIELDS
-                    if key in response['result']
-                }
-                return lite_metadata
-            else:
-                return {
-                    'error': response['error'],
-                    'identifier': ocaid,
-                }
-        except HTTPError:
-            logger.warning(f'Error fetching metadata for {ocaid}')
-            return None
-
-    async def cache_ia_metadata(self, ocaids: list[str]):
-        invalid_ocaids = {ocaid for ocaid in ocaids if not is_valid_ocaid(ocaid)}
-        if invalid_ocaids:
-            logger.warning(f"Trying to cache invalid OCAIDs: {invalid_ocaids}")
-        valid_ocaids = list(set(ocaids) - invalid_ocaids)
-        batches = list(batch_until_len(valid_ocaids, 3000))
-        # Start them all async
-        tasks = [asyncio.create_task(self._get_lite_metadata(b)) for b in batches]
-        for task in tasks:
-            for doc in await task:
-                self.ia_cache[doc['identifier']] = doc
-
-        missing_ocaids = [ocaid for ocaid in valid_ocaids if ocaid not in self.ia_cache]
-        missing_ocaid_batches = list(batch(missing_ocaids, 6))
-        for ocaids in missing_ocaid_batches:
-            # Start them all async
-            tasks = [
-                asyncio.create_task(self._get_lite_metadata_direct(ocaid))
-                for ocaid in ocaids
-            ]
-            for task in tasks:
-                lite_metadata = await task
-                if lite_metadata:
-                    self.ia_cache[lite_metadata['identifier']] = lite_metadata
-
     def cache_edition_works(self, lo_key, hi_key):
         q = """
             SELECT works."Key", works."JSON"
@@ -376,7 +196,7 @@ class LocalPostgresDataProvider(DataProvider):
 
     async def cache_cached_editions_ia_metadata(self):
         ocaids = list({doc['ocaid'] for doc in self.cache.values() if 'ocaid' in doc})
-        await self.cache_ia_metadata(ocaids)
+        await self.preload_metadata(ocaids)
 
     def find_redirects(self, key):
         """Returns keys of all things which redirect to this one."""
@@ -417,16 +237,6 @@ class LocalPostgresDataProvider(DataProvider):
         else:
             return self.get_editions_of_work_direct(work)
 
-    def get_metadata(self, identifier):
-        if identifier in self.ia_cache:
-            logger.debug("IA metadata cache hit")
-            return self.ia_cache[identifier]
-        elif not is_valid_ocaid(identifier):
-            return None
-        else:
-            logger.debug("IA metadata cache miss")
-            return ia.get_metadata_direct(identifier)
-
     def get_document(self, key):
         if key in self.cache:
             logger.debug("get_document cache hit %s", key)
@@ -446,9 +256,9 @@ class LocalPostgresDataProvider(DataProvider):
             return row[0]
 
     def clear_cache(self):
+        super().clear_cache()
         self.cached_work_editions_ranges.clear()
         self.cache.clear()
-        self.ia_cache.clear()
 
 
 def simple_timeit(fn):
